+++ CHANGELOG.md
@@ -9,6 +9,7 @@ No date
 28.01.2023
 
 - `--cache-downloads` option. In 28.0, `mist` cached downloads into `/private/tmp/com.ninxsoft.mist`. This is now optional.
+- New experimental `--set-securebootlevel` option (in `--erase` mode only) uses the command `bputil -f -u $current_user -p $account_password` to ensure that the OS is reset to a high secure boot level after reinstallation (thanks to @mvught). 
 
 Bugfixes:
 - `mist` result is now correctly interpreted when checking for a newer version.